Permission clean up: use noteOp where necessary
Ensure we call noteOp for the relevant app-op at least once for every
operation in MediaProvider, be it through FUSE or ContentResolver.

Refactor some of the methods in PermissionUtils:
* Distinguish methods that call noteOp from those that call checkOp
* Distinguish methods that check permission then app-op from those that
  do the opposite
* Prepare methods to take in (noteOp) message parameter

Sets an app-op description for each FUSE LocalCallingIdentity so that it
can be used as an app-op message when calling noteOp.

Fix: 149473595
Test: atest FuseDaemonHostTest MediaProviderTests CtsProviderTestCases
Change-Id: If18e2f8d53e617e0c3d58133cd1812874b2d5554