Fix wrong initialization order in FAdviser
FAdviser declares thread_ before mutex_. This can cause a race leading
to crash when we start the thread_ (in the initializer list) and the
new thread tries to acquire mutex_ before it is initialized.

Now, we declare mutex_ before thread_ ensuring that mutex_ will always
be initialized before thread_ runs.

Test: m
Bug: 154416156
Change-Id: I9a6ab9bf744c2158fdf14a03c70fe95cf9ad61bc