Update media_type as NONE for hidden files created via filepath.
Hidden files and files inside hidden directory or directory containing
.nomedia file are consider hidden, and should be updated with media type
as MEDIA_TYPE_NONE.

When a file created or renamed, the new file might be a hidden file or
inside a hidden directory, and this hidden file must be updated with
media_type as MEDIA_TYPE_NONE. We achieve the same in this CL by
scanning the newly created or renamed file.

Added tests for creating, renaming and deleting hidden file.

Test: atest FuseDaemonHostTest
Test: atest packages/providers/MediaProvider
Bug: 148585977, 150927291
Change-Id: If6fd1b950a46ea7b43953661e7fbdaaf126bf3ee