Give up trying to give a detailed message in noteOp
We used to associate an app-op message (describing the op) to a
LocalCallingIdentity. This can lead to mix ups when the same package
uses file path access on multiple threads, in addition, noteOp results
are cached, which means we can't always provide a correct detailed
message.
This CL provides a more generic message: either "via MediaProvider" if
the current operation started from Binder, or "via FUSE" if it came from
FUSE.

Test: atest FuseDaemonHostTest MediaProviderTests CtsProviderTestCases
Bug: 149473595
Change-Id: I3d5567e7a57ceaddb78bac085ba4d7acc9271284