Get ourselves some Error Prone.
As we move towards becoming a Mainline module, it's important that
we catch as many bugs as possible during initial code review, and
Error Prone is designed to do exactly that.

This change fixes all outstanding Error Prone issues, and upgrades
several of them from "warnings" to "errors" to ensure we don't
regress again in the future.

Bug: 142561358
Test: ./build/soong/soong_ui.bash --make-mode MediaProviderLegacy com.google.android.mediaprovider MediaProviderTests RUN_ERROR_PRONE=true
Change-Id: I18a94a70d14f65da2af918ac40c83b2e8de97291