Use stat.st_dev for checking if two paths point at different filesystems
Comparing statvfs.f_fsid might be insufficient because
fsid is set to 0 on some filesystems.
For example, on ARC, fsid is 0 for both /storage (tmpfs)
and /storage/emulated/0 (FUSE).

stat.st_dev is device ID, so may not be persistent across
reboots but unique for comparison at runtime,
which is good enough for this purpose.

Bug: 157383101
Test: tast run eve-arc-r arc.DataCollector.vm
Change-Id: I220d9a6271a2234a8afe56f5a86d4f725c3963fe