Always update quota type when handling an update.
Apps can insert rows in the MediaStore database without the file
existing yet; in that case, we didn't update the quota type in the
update handler either, because the media type of the file hasn't changed
in-between. To fix those, update the quota type on the filesystem
regardless of whether the media type has actually changed.

Also, don't try to set quota type on a file that doesn't exist yet.

Bug: 156336269
Test: inspect quota on filesystem
Change-Id: I552d5f9d9e5a72323d1076be305cef1a87c5791e