Gracefully handle trashing/pending missing files.
We've heard of some odd conditions where the underlying file has
gone missing before a media scan can update the database.  The best
we can do is let the rename of the database row proceed, and the
next scan operation will likely reconcile the database against the
real filesystem contents.  (It's not the fault of the calling app,
so it's not fair to throw the exception back at them.)

Bug: 155765090
Test: atest --test-mapping packages/providers/MediaProvider
Change-Id: I4bbe1369a64843950d9a13b1b533b713d7ac9457