Use volume names instead of "Unknown".
Partners have identified that when the user has placed media files
at the top-level of multiple attached devices, then we surface
duplicate buckets called "Unknown" which can be confusing.

This change adjusts that behavior to use the volume name of each
secondary SD card as the bucket name to reduce that confusion.  We
continue to have a single "Unknown" bucket for the primary storage
device hosted at VOLUME_EXTERNAL_PRIMARY.

Bug: 158144861
Test: atest --test-mapping packages/providers/MediaProvider
Change-Id: I1a944b6b845b0f68f97a4209b7603d55aa57de7b