MediaProvider: Change owners on legacy app creates
If a legacy app creates a file at a path that has an entry in the files db
with the same path (_data), we transfer ownership to the legacy app to
prevent the original owner from being able to observe the contents. Note
that creates from non-legacy apps will fail.

Test: atest LegacyStorageHostTest
Bug: 150507736
Change-Id: Ic6186c3e84a09222db09300b8baec0e698e1e0ba