Add pulled metrics for external storage accesses
Start collecting stats regarding the number of accesses to external
storage and the accesses mime types. This includes all successful and
failed access attempts through both ContentResolver and FUSE.

Each UID will have its own separate stats and a separate atom will be
pulled for it. We don't send all of the UID stats since we don't want to
go over the budget, so we're limited by a PulledMetrics#SAMPLES_LIMIT.

NOTE: We don't log accesses done by MediaProvider.

NOTE: This is a manual merge since isDirectoryCreationOrDeletionAllowedForFuse was removed

Test: make statsd_testdrive ; ./out/host/linux-x86/bin/statsd_testdrive \
      -p com.google.android.providers.media.module 10085
Bug: 149669087

Change-Id: I59f9f42c9da4b17299a681d17e91859d6389ccd9