Don't call into DownloadManager as part of a FUSE operation.
Some ramdumps showed that android.process.media was stuck on an unlink()
call into FUSE. The FUSE thread handling the unlink() was making a
binder call back into android.process.media, causing a deadlock. A
likely candidate for that call is
DownloadManager.onMediaStoreDownloadsDeleted(), since it's called a part
of unlink(), and DownloadManager does live in android.process.media.
Move it to a background thread, since it's not critical for this to be
handled at the same time as the delete itself.

Bug: 158657599
Test: atest CtsScopedStorageHostTest
Change-Id: Ia5495c0065e64a37c75c182d3ad23d0449759218