Performance optimization for FUSE queries.
When returning Cursor objects to remote apps, we need to configure
the notification Uri so they can observe changes.  However, when we
know that a Cursor will only be used briefly in-process and not
passed across Binder, we can skip this registration and avoid two
Binder calls.

We already had this optimization for ModernMediaScanner, and this
change expands that logic to also cover queries run by FUSE threads.

Bug: 159446622
Test: atest --test-mapping packages/providers/MediaProvider
Change-Id: Id2373a97c49804ea2cf04962d7d81119a32ab430