Enable fuse passthrough mode in FuseDaemon
This patch enables the FUSE passthrough feature in the MediaProvider.
FUSE passthrough is activated only if "persist.sys.fuse.passthrough" is
set to true.  No further special conditions of the file are required for
the passthrough feature to be enabled.

FUSE passthrough is a feature that allows creating at OPEN/CREATE time a
direct channel from the process accessing the FUSE file system to the
file that lies in the lower file system. All the following read/write
operations targeting a file that was opened with FUSE passthrough mode
on target the lower file system directly, without the overhead of
traversing the FUSE daemon.

  go/fuse-passthrough

Bug: 168023149
Test: atest ScopedStorageTest
Signed-off-by: Alessio Balsini <balsini@google.com>
Change-Id: I3a4be233d0c8b1566b4eab72cf3be30b8d30353f