FuseDaemon: remove .flush callback
Our implementation of the flush file operation does nothing except for
tracing.  The result is that the FUSE module in the kernel still
forwards the flush requests to the FUSE daemon.
Remove the flush implementation, so that the kernel can handle the flush
internally (doing nothing), instead of forwarding the flush request to
the FUSE daemon.

Test: N/A, this function does not change the system behavior
Signed-off-by: Zimuzo Ezeozue <zezeozue@google.com>
Signed-off-by: Alessio Balsini <balsini@google.com>
Change-Id: I12f917e0833fa940ad5fb84608be0434b232b047