Return the correct error code for cross-user lookups
EACCES: “Permission denied.” The file permissions do not allow the
attempted operation.

EPERM: “Operation not permitted.” Only the owner of the file (or other
resource) or processes with special privileges can perform the
operation.

EACCESS should be used in this case as the file permissions do not allow
cross-user lookups.

Bug: 161768276
Test: Manually using the debug app attached in b/161768276#comment5
Change-Id: Ic44d8170a3fdbd07d122929a07d9caafa7c7d57d