Delete cached transcode files on deleting original file
When legacy apps try to read a video file in modern format, we transcode
the file and return the FileDescriptor of transcode file. This transcode
file is cached and will be used when same app or a different legacy app
tries to open this file. The cached transcode files will not be directly
accessible to apps hence when the original file video file in modern
format is deleted, we want to delete the cached transcode file too.

Chanaged MediaProvider#delete() to delete the transcode file when the
original file is deleted.

For apps that bypass database operations, we don't delete the transcode
files in filePath operation. However these apps are expected to delete
database row on their own, or call scanFile() which will delete the
database row and also delete the transcode file.

Bug: 158466177
Test: N.A.
Change-Id: I124e58e5f7d6f577f805e74437da1379ee25c5ec