Use synthetic external view for playlist member uri
This allows audio which is saved on secondary external storage
can be added into a playlist which is saved on primary externals
storage. Without this fix, the check for calling permissions would
fail as the Audio may not exist on the same volume as the playlist.

Bug: 165026333
Test: atest MediaProviderTests
Change-Id: Iaf3563df542bfe27a61d46e6041bb0a3afd69604
(cherry picked from commit 831806db2efbc36121ec7e3ae80767dbe24f7645)