Prevent unnecessary node invalidation
Previously, in an attempt to fix case-insensitivity bugs, we
aggressively invalidated nodes anytime we received a lookup for an
existing node. This could degrade performance unnecessarily because
whenever we readdir(2) we lookup child nodes and invalidate any
existing nodes.

Now, we only invalidate existing nodes if there has been any history
of a case insensitive match.

Bug: 168666994
Test: atest ScopedStorageHostTest#testCaseInsensitivity
Change-Id: I04af224f2e15db9be6c19c4193577d4b7156d405