Solve playlist items deletion issue
Deleting playlist items calls into resolvePlaylistIndex, which returns
-1 if multiple items in the playlist match. In return,
removePlaylistMembers deletes the entire content of the playlist in that
case. This breaks deleting multiple items from the playlist, as it will
always result in deleting all of the items.

This CL fixes that by introducing resolvePlaylistIndexes which returns
an array of indexes so that removePlaylistMembers can delete all of
them.

Bug: 162564762
Test: atest PlaylistTest
Change-Id: I190d0ba1094f2733335521d19cd40ee9e4e9bc62