Add static allow list and transcode list
Some apps might run into error when MediaProvider is transcoding the
modern file. To handle these app errors for team food, we allow some
apps to bypass transcoding. The allow list is a static list for now,
and will be populated based on the server list in the future.

Transcode list will force transcoding for thei given packages. This list
will be removed when we get information about app capabilities from the
manifest file.

Bug: 169327180
Test: Manual
Change-Id: I940912658a9275c2dcb487a6b37081b13e4fb68e