Don't keep a calling identity cache for other users.
When MediaProvider user 0 is also handling filesystem requests for other
users (which can happen in "app cloning" designs), the calling identity
cache that we're using can't be kept up-to-date correctly, because we
can't get permission / app-op events for other users. So, do not cache
in case the userId of the caller does not match the userId of the
MediaProvider mainline module.

Bug: 168281115
Test: TBD
Change-Id: I714615350a31b6cac8f7d051a3cb946388359719