Disable dentry cache for clone app users.
If a file manager running in user 0 deletes a file from a clone app
user, the clone app user FUSE daemon may not be aware of it, because it
has its own dentry cache that thinks the file still exists. To prevent
problems like this, disable the dentry cache for clone app users, such
that they will always see the latest state on the filesystem.

Bug: 161768276
Test: atest CtsScopedStorageTest
Change-Id: I96550642a0d27b30f35edc7177ba90ce21c84eae