Transcode only for HEVC video codec format
Added a check in TranscodeHelper#shouldTranscode to check the video
codec format of the given video file to make the decision to transcode
or not. This could slow down the lookups because of a db query. However,
it only affects mp4 files in DCIM/Camera directory.

With the current change, node will still have FLAG_TRANSFORM_TRANSCODING
for all video files in DCIM/Camera, but FLAG_TRANSFORM_SUPPORTED will
only be set if the video codec is HEVC and if the app needs seamless
transcoding.

Also fixed shouldTranscode to get package name from the given uid. We
don't set LocalCallingIdentity before calling shouldTranscode, hence
getSharedPackages will not be able to get calling package name.

Bug: 169849764
Test: Manual, tested that transcoding is initiated only for video files
with video codec format as HEVC.

Change-Id: I8876075d4bd3749dc839def2c4b31e0b7d5aa85f