Fix transcoding logs and toasts
Previously we showed toast when a transcoding 'wait' started, this was
misleading because several IO requests could 'wait' for the same
transcoding job.

Now we log separately for the wait and actual transcode
events. Additionally, we only show toast for the actual transcode
events. In the process, refactored the logging code slightly for
improved readability.

Fixed bug where there was a transcode cache hit and we still attempted
transcode

Test: Manual
Bug: 170660980
Change-Id: If9ed533d70cc749742253b6fb1399604dac2d7ac